requirements.txt
pandas
pillow
tqdm
numpy
torch 
torchvision 
torchaudio
lightning
matplotlib
scipy
transformers
sentence-transformers
openpyxl
mlflow
mlflow_export_import
tabulate
boto3
parsl